@@ -5,13 +5,13 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.2
-  intl: ^0.16.1
+  http: "0.13.0"
+  intl: "0.17.0"
 
 dev_dependencies:
   flutter_test:
