@@ -4,11 +4,11 @@ import 'package:wui/widgets/button.widget.dart';
 class WuiDialog extends StatelessWidget { 
 
   static Future<dynamic> open(BuildContext context, {
-    @required dynamic title,
-    @required dynamic message,
-    @required List<dynamic> buttons,
-    int defaultAction,
-    EdgeInsets insetPadding,
+    required dynamic title,
+    required dynamic message,
+    required List<dynamic> buttons,
+    int? defaultAction,
+    EdgeInsets? insetPadding,
   }) async {
     
     return await showDialog(
@@ -17,15 +17,13 @@ class WuiDialog extends StatelessWidget {
       builder: (context) => AlertDialog(
         title: title is String ? Text(title) : title,
         content: message is String ? Text(message) : message,
-        actions: buttons.asMap().map((index, item) {
-          Widget widget;
+        actions: List<Widget>.from(buttons.asMap().map((index, item) {
+          Widget? widget;
           
           if(buttons[index] is String) {
-            widget = FlatButton(
-              padding: EdgeInsets.symmetric(horizontal: 14, vertical: 0),
-              textTheme: defaultAction == index ? ButtonTextTheme.primary : ButtonTextTheme.normal,
+            widget = TextButton(
               child: DefaultTextStyle(
-                style: Theme.of(context).textTheme.bodyText1,
+                style: Theme.of(context).textTheme.bodyText1!,
                 child: Text(buttons[index])
               ),
               onPressed: () {
@@ -41,22 +39,22 @@ class WuiDialog extends StatelessWidget {
                   textTheme: defaultAction == index ? ButtonTextTheme.primary : ButtonTextTheme.normal,
                 ),
               ),
-              child: widget
+              child: widget!
             );
           }
 
           return MapEntry(index, widget);
-        }).values.toList()
+        }).values.toList())
       )
     );
 
   }
 
   final dynamic title;
-  final Widget child;
-  final List<dynamic> actions;
-  final int defaultActions;
-  final Function(int) onPressed;
+  final Widget? child;
+  final List<dynamic>? actions;
+  final int? defaultActions;
+  final Function(int)? onPressed;
 
   WuiDialog({
     this.title,
@@ -78,7 +76,7 @@ class WuiDialog extends StatelessWidget {
           ...[Padding(
             padding: const EdgeInsets.fromLTRB(24, 24, 24, 8),
             child: DefaultTextStyle(
-              style: Theme.of(context).textTheme.bodyText2.copyWith(
+              style: Theme.of(context).textTheme.bodyText2!.copyWith(
                 fontSize: 20, 
                 fontWeight: FontWeight.w500,
                 color: Colors.black87
@@ -97,21 +95,21 @@ class WuiDialog extends StatelessWidget {
             padding: const EdgeInsets.all(16),
             child: Row(
               mainAxisAlignment: MainAxisAlignment.end,
-              children: actions.asMap().map((index, item) {
-                Widget widget;
-                if(actions[index] is String) {
+              children: actions!.asMap().map((index, item) {
+                Widget? widget;
+                if(actions![index] is String) {
                   widget = WuiButton(
                     size: WuiButtonSize.small,
                     isDense: true,
                     theme: defaultActions != null ? (defaultActions == index ? WuiButtonTheme.transparentPrimary : WuiButtonTheme.transparent) : WuiButtonTheme.transparent,
-                    child: Text(actions[index]),
+                    child: Text(actions![index]),
                     onPressed: () {
-                      onPressed(index);
+                      onPressed!(index);
                     },
                   );
                 }
                 return MapEntry(index, widget);
-              }).values.toList()
+              }).values.toList() as List<Widget>
             ),
           )]
         ],
