@@ -1,20 +1,24 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:wui/colors/primary.color.dart';
 
 class WuiSliverAppBar extends StatefulWidget {
 
-  final List<Widget> actions;
-  final Widget title;
-  final Widget subtitle;
-  final ScrollController controller;
-  final BuildContext context;
+  final List<Widget>? actions;
+  final Widget? floatingTitle;
+  final Widget? title;
+  final Widget? subtitle;
+  final PreferredSizeWidget? bottom;
+  final bool showExpanded;
 
   WuiSliverAppBar({
-    @required this.controller,
-    @required this.context,
+    this.floatingTitle,
     this.title,
     this.subtitle,
     this.actions,
+    this.bottom,
+    this.showExpanded = true
   });
 
   @override
@@ -23,8 +27,12 @@ class WuiSliverAppBar extends StatefulWidget {
 
 class _WuiSliverAppBarState extends State<WuiSliverAppBar> with SingleTickerProviderStateMixin {
 
-  AnimationController _animationController;
+  late AnimationController _animationController;
   double _flexOpacity = 1;
+  Timer? _sizeChangeDebounce;
+  GlobalKey _sizeChangeKey = GlobalKey();
+  double _maxHeight = 0;
+  double _appBarHeight = 0;
 
   @override
   void initState() {
@@ -33,17 +41,9 @@ class _WuiSliverAppBarState extends State<WuiSliverAppBar> with SingleTickerProv
       duration: const Duration(milliseconds: 200),
       animationBehavior: AnimationBehavior.normal
     );
-    widget.controller.addListener(() { 
-      if(this.mounted) {
-        double max = ((MediaQuery.of(widget.context).size.height / 2) - 56) - 32;
-        if(widget.controller.offset > max) {
-          setState(() { _flexOpacity = 0; });
-          _animationController.forward();
-        } else {
-          setState(() { _flexOpacity = (max - widget.controller.offset) / max; });
-          _animationController.reverse();
-        }
-      }
+    WidgetsBinding.instance!.addPostFrameCallback((timeStamp) {
+      _appBarHeight = 56 + (widget.bottom?.preferredSize.height ?? 0) + MediaQuery.of(context).padding.top;
+      _maxHeight = _sizeChangeKey.currentContext!.size!.height - _appBarHeight;
     });
     super.initState();
   }
@@ -53,7 +53,7 @@ class _WuiSliverAppBarState extends State<WuiSliverAppBar> with SingleTickerProv
     return SliverAppBar(
       backgroundColor: surfaceColor,
       actions: widget.actions,
-      title: AnimatedBuilder(
+      title: widget.floatingTitle != null ? widget.floatingTitle : AnimatedBuilder(
         animation: _animationController,
         builder: (context, child) {
           return Opacity(
@@ -67,40 +67,64 @@ class _WuiSliverAppBarState extends State<WuiSliverAppBar> with SingleTickerProv
       ),
       floating: false,
       pinned: true,
-      expandedHeight: (MediaQuery.of(context).size.height / 2),
-      flexibleSpace: Container(
-        padding: EdgeInsets.only(top: 56),
-        child: Center(
-          child: Opacity(
-            opacity: _flexOpacity,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                DefaultTextStyle(
-                  style: Theme.of(context).textTheme.headline4.copyWith(
-                    fontSize: 24,
-                    fontWeight: FontWeight.w500
-                  ), 
-                  child: Container(
-                    constraints: BoxConstraints(
-                      maxWidth: 340
+      expandedHeight: widget.showExpanded ? MediaQuery.of(context).size.height / 2 : 0,
+      bottom: widget.bottom,
+      flexibleSpace: NotificationListener(
+        onNotification: (SizeChangedLayoutNotification notification) {
+          double currentHeight = _sizeChangeKey.currentContext!.size!.height - _appBarHeight;
+          double opacity = (currentHeight / _maxHeight);
+          if(_sizeChangeDebounce != null) _sizeChangeDebounce!.cancel();
+          _sizeChangeDebounce = Timer(Duration.zero, () {
+              setState(() { _flexOpacity = opacity; });
+              if(_flexOpacity == 0) {
+                _animationController.forward();
+                return;
+              }
+              _animationController.reverse();
+          });
+          return true;
+        },
+        child: Opacity(
+          opacity: _flexOpacity,
+          child: SizeChangedLayoutNotifier(
+            key: _sizeChangeKey,
+            child: Container(
+              padding: EdgeInsets.only(top: 56),
+              child: Center(
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    DefaultTextStyle(
+                      style: Theme.of(context).textTheme.headline4!.copyWith(
+                        fontSize: 28,
+                        fontWeight: FontWeight.w500
+                      ), 
+                      child: Container(
+                        constraints: BoxConstraints(
+                          maxWidth: 340
+                        ),
+                        child: widget.title
+                      )
                     ),
-                    child: widget.title
-                  )
-                ),
-                DefaultTextStyle(
-                  style: Theme.of(context).textTheme.bodyText2, 
-                  child: Container(
-                    constraints: BoxConstraints(
-                      maxWidth: 340
+                    SizedBox(height: 8),
+                    DefaultTextStyle(
+                      textAlign: TextAlign.center,
+                      style: Theme.of(context).textTheme.bodyText2!.copyWith(
+                        fontSize: 16
+                      ), 
+                      child: Container(
+                        constraints: BoxConstraints(
+                          maxWidth: 300
+                        ),
+                        child: widget.subtitle
+                      )
                     ),
-                    child: widget.subtitle
-                  )
-                ),
-              ],
+                  ],
+                )
+              )
             ),
           )
-        )
+        ),
       ),
     );
   }
