@@ -3,8 +3,8 @@ import 'package:wui/widgets/list-tile.widget.dart';
 
 class WuiSliverList extends StatelessWidget {
 
-  final Function builder;
-  final int itemCount;
+  final Function? builder;
+  final int? itemCount;
   WuiSliverList({
     this.builder,
     this.itemCount
@@ -15,10 +15,10 @@ class WuiSliverList extends StatelessWidget {
     return SliverList(
       delegate: SliverChildBuilderDelegate(
         (_, index) {
-          Widget item = builder(context, index);
+          Widget? item = builder!(context, index);
           return WuiListTileTheme(
             borderMode: WuiListTileBorderMode.full,
-            divider: (index < itemCount - 1 ? true : false),
+            divider: (index < itemCount! - 1 ? true : false),
             child: item
           );
         },
