@@ -6,15 +6,15 @@ class WuiSnackBar {
     dynamic content,
     bool rootNavigator = false
   }) {
-    Scaffold.of(context).hideCurrentSnackBar();
-    Scaffold.of(context).showSnackBar(SnackBar(
+    ScaffoldMessenger.of(context).hideCurrentSnackBar();
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: content is Widget ? content: Text(content),
       behavior: SnackBarBehavior.fixed,
       action: SnackBarAction(
         label: "DISMISS",
         textColor: Colors.white,
         onPressed: () {
-          Scaffold.of(context).hideCurrentSnackBar();
+          ScaffoldMessenger.of(context).hideCurrentSnackBar();
         },
       )
     ));
