@@ -6,13 +6,11 @@ Color surfaceColor = Colors.white;
 class WuiThemeData {
 
   static light(BuildContext context, {
-    Colors primarySwatch,
+    Color? primarySwatch,
   }) {
     return ThemeData(
-      primarySwatch: primarySwatch ?? Colors.deepPurple,
+      primarySwatch: primarySwatch as MaterialColor? ?? Colors.deepPurple,
       scaffoldBackgroundColor: surfaceColor,
-      canvasColor: Colors.white,
-
       appBarTheme: AppBarTheme(
         iconTheme: IconThemeData(
           color: textColor
@@ -23,7 +21,7 @@ class WuiThemeData {
         color: Colors.white,
         elevation: 4,
         textTheme: TextTheme(
-          headline6: Theme.of(context).textTheme.headline6.copyWith(
+          headline6: Theme.of(context).textTheme.headline6!.copyWith(
             color: textColor
           )
         )
@@ -35,26 +33,26 @@ class WuiThemeData {
         color: textColor
       ),
       textTheme: TextTheme(
-        caption: Theme.of(context).textTheme.caption.copyWith(
+        caption: Theme.of(context).textTheme.caption!.copyWith(
           color: textColor
         ),
-        subtitle1: Theme.of(context).textTheme.subtitle1.copyWith(
+        subtitle1: Theme.of(context).textTheme.subtitle1!.copyWith(
           color: textColor
         ),
-        subtitle2: Theme.of(context).textTheme.subtitle2.copyWith(
+        subtitle2: Theme.of(context).textTheme.subtitle2!.copyWith(
           color: textColor,
           fontSize: 16
         ),
-        headline4: Theme.of(context).textTheme.headline4.copyWith(
+        headline4: Theme.of(context).textTheme.headline4!.copyWith(
           color: textColor
         ),
-        headline6: Theme.of(context).textTheme.headline6.copyWith(
+        headline6: Theme.of(context).textTheme.headline6!.copyWith(
           color: textColor
         ),
-        bodyText1: Theme.of(context).textTheme.bodyText1.copyWith(
+        bodyText1: Theme.of(context).textTheme.bodyText1!.copyWith(
           color: textColor
         ),
-        bodyText2: Theme.of(context).textTheme.bodyText2.copyWith(
+        bodyText2: Theme.of(context).textTheme.bodyText2!.copyWith(
           color: textColor
         ),
       )
